+++ config/routes.rb
@@ -1,5 +1,5 @@
 GovukPublishingComponents::Engine.routes.draw do
-  root :to => 'component_guide#index', as: :component_guide
+  root to: 'component_guide#index', as: :component_guide
   get ':component/preview' => 'component_guide#preview', as: :component_preview_all
   get ':component/:example/preview' => 'component_guide#preview', as: :component_preview
   get ':component' => 'component_guide#show', as: :component_doc